+++ lighthouse-cli/performance-experiment/formatters/config-panel.js
@@ -1,0 +1,37 @@
+/**
+ * @license
+ * Copyright 2016 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+const path = require('path');
+const fs = require('fs');
+const Formatter = require('./perf-x-formatter');
+const html = fs.readFileSync(path.join(__dirname, 'partials/config-panel.html'), 'utf8');
+
+class ConfigPanel extends Formatter {
+  static getFormatter(type) {
+    switch (type) {
+      case 'html':
+        return html;
+
+      default:
+        throw new Error('Unknown formatter type');
+    }
+  }
+}
+
+module.exports = ConfigPanel;
