+++ lighthouse-cli/performance-experiment/formatters/partials/config-panel.html
